requirements.txt
six
pyyaml
PyInstaller

# for reading xlsx files
# openpyxl

# for writing xlsx files
# xlsxwriter
